% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_cor_heatmap_pheno}
\alias{plot_cor_heatmap_pheno}
\title{plot_cor_heatmap_pheno}
\usage{
plot_cor_heatmap_pheno(var1, var2, method = "pearson")
}
\arguments{
\item{pheno_int}{}
}
\description{
plot_cor_heatmap_pheno
}
