% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_group_box}
\alias{plot_group_box}
\title{plot_group_box}
\usage{
plot_group_box(
  omic_df,
  pheno_df,
  gene = "ACVR1B",
  groups = "group",
  comparisons = list(c("Base", "Day1"), c("Day1", "Day90"), c("Base", "Day90")),
  test = "ttest"
)
}
\arguments{
\item{test}{}
}
\description{
plot_group_box
}
