% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyomics.R
\name{tidyomics}
\alias{tidyomics}
\title{tidyomics}
\usage{
tidyomics()
}
\value{
No value is returned
}
\description{
A helper package to install and attach the core packages of the tidyomics ecosystem.
}
\examples{
\dontrun{
# Install the core packages of the tidyomics ecosystem
BiocManager::install("tidyomics")

# Attach the core packages of the tidyomics ecosystem
library(tidyomics)
}
}
