% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{tidyomics_deps}
\alias{tidyomics_deps}
\title{List all tidyomics dependencies}
\usage{
tidyomics_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidyomics packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
List all tidyomics dependencies
}
