% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyomics_packages}
\alias{tidyomics_packages}
\title{List all packages loaded by the tidyomics metapackage}
\usage{
tidyomics_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tidyomics in the list?}
}
\value{
All packages loaded by the tidyomics metapackage.
}
\description{
List all packages loaded by the tidyomics metapackage
}
\examples{
tidyomics_packages()

}
