% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{tidyomics_sitrep}
\alias{tidyomics_sitrep}
\title{Get a situation report on the tidyomics}
\usage{
tidyomics_sitrep()
}
\description{
This function gives a quick overview of the versions of R and RStudio as
well as all tidyomics packages. It's primarily designed to help you get
a quick idea of what's going on when you're helping someone else debug
a problem.
}
\examples{
\dontrun{
tidyomics_sitrep()
}

}
