% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\docType{package}
\name{tidyomics-package}
\alias{tidyomics}
\alias{tidyomics-package}
\title{tidyomics: Easily Install and Load the 'tidyomics'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The tidyomics ecosystem is a set of packages for ’omic data analysis that work together in harmony; they share common data representations and API design, consistent with the tidyverse ecosystem. The tidyomics package is designed to make it easy to install and load core packages from the tidyomics ecosystem with a single command.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyomics/tidyomics}
  \item Report bugs at \url{https://github.com/tidyomics/tidyomics/issues}
}

}
\author{
\strong{Maintainer}: Stefano Mangiola \email{mangiola.s@wehi.edu.au} (\href{https://orcid.org/0000-0001-7474-836X}{ORCID})

Authors:
\itemize{
  \item Michael Love \email{michaelisaiahlove@gmail.com} (\href{https://orcid.org/0000-0001-8401-0545}{ORCID})
}

}
\keyword{internal}
